handle different package lists
RHEL 7-based systems do not have Python 3.5 as part of the default
installation, which is a requirement for this install script. In
order to get this package, one needs to install from SCL or from
source. Consequently, it isn't absolutely necessary (at least at this
point) to require the related `-devel` package in the install script.

This should be noted that this is really a stop-gap, and a more
robust solution is needed.